+++ src/components/Chart.vue
@@ -3,7 +3,7 @@
     <ChartOptions :crypto="crypto" :base="base" v-model="base" />
     <div class="stats">
       <span class="price">
-        <span class="h2">$95.23</span>
+        <h2>$95.23</h2>
       </span>
       <span class="statsgroup">
         <span>24h Change</span>
@@ -19,63 +19,36 @@
       </span>
     </div>
     <div class="chart-container">
-      <VChart :option="option" />
+      <CandlestickChart :crypto="crypto" :base="base" />
     </div>
   </section>
 </template>
 
-
-<script>
+<script setup>
 import ChartOptions from "@/components/ChartOptions.vue";
-import { onMounted, reactive, ref } from "vue";
-import options from "@/config/chartOptions.js";
+import CandlestickChart from "@/components/CandlestickChart.vue";
+import { ref } from "vue";
 
-import { use } from "echarts/core";
-import { CanvasRenderer } from "echarts/renderers";
-import { CandlestickChart } from "echarts/charts";
-import { GridComponent, DataZoomComponent, TooltipComponent } from "echarts/components";
-import VChart from "vue-echarts";
-
-export default {
-  props: {
-    crypto: {
-      type: String,
-      required: true
-    },
-
-    base: {
-      type: String,
-      required: true
-    }
+const props = defineProps({
+  crypto: {
+    type: String,
+    required: true
   },
 
-  components: { VChart, ChartOptions },
-
-  setup(props) {
-    use([CandlestickChart, CanvasRenderer, GridComponent, DataZoomComponent, TooltipComponent ]);
-
-    const { crypto } = props;
-    const base = ref(props.base);
-    const option = reactive(options);
-
-    // Temporary load the data
-    onMounted(() => {
-      fetch("https://api.binance.com/api/v3/klines?symbol=ETHEUR&interval=5m")
-        .then(res => res.json())
-        .then(data => {
-          data = data.slice(-100, -1);
-          // Convert to float
-          data = data.map(candle => candle.slice(0, 5).map(c => parseFloat(c)));
-          option.series.data = data;
-      });
-    });
-
-    return { crypto, base, option }
+  base: {
+    type: String,
+    required: true
   }
-}
+});
+
+const { crypto } = props;
+const base = ref(props.base);
 </script>
 
-<style lang="sass">
+<style lang="sass" scoped>
+  h2
+    font-weight: normal
+    
   .chart-container
     height: 70vh
     width: 100%
@@ -97,15 +70,10 @@ export default {
     display: inline-flex
     flex-direction: column
     justify-content: center
-    
-  canvas
-    border-radius: 0 0 1.5em 1.5em
-    cursor: crosshair
 
   .green
     color: $green
 
   .red
     color: $red
-
 </style>
