+++ src/config/candlestickChartOptions.js
